refactor(pipeline): extract prompt data functions to prompt_data.go [#466]
Move buildPromptData, computeDiffContext, computePlanHash, extractPRURL,
and defaultMaxDiffBytes const from engine.go to prompt_data.go.

Remove encoding/json import from engine.go (only used by extractPRURL).
